some update

+++ core/src/mwin/winlib/combobox.c
@@ -37,6 +37,7 @@
 #include "windows.h"
 #include "windowsx.h"
 #include "mwsystem.h"
+#include "wintools.h"
 
 #define WinMalloc(n)	malloc((n))
 #define WinFree(p)	free(p)
@@ -363,7 +364,7 @@ DefComboboxProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
         lp->bExtended  = TRUE;
         lp->hFont = 0;
         lp->hWndParent = lpcs->hwndParent;
-        lp->nID  = (UINT)lpcs->hMenu;
+        lp->nID  = (UINT)(intptr_t)lpcs->hMenu;
 
         dwStyle = GetWindowLong(hWnd, GWL_STYLE);
         dwStyle &= ~(WS_VSCROLL | WS_HSCROLL | WS_BORDER | WS_DLGFRAME | WS_THICKFRAME);
